latitude and longitude now display

+++ backend/src/main/java/edu/lehigh/cse216/kta221/backend/MessageRequest.java
@@ -25,7 +25,7 @@ public class MessageRequest {
 
     public String file;
 
-    public long longitude;
+    public float longitude;
 
-    public long latitude;
+    public float latitude;
 }
